Rename request to mappedparameter and make it consistent
