+++ src/Dependency/Service/DependencyService.php
@@ -19,9 +19,9 @@ namespace Pimcore\Bundle\StudioBackendBundle\Dependency\Service;
 use Pimcore\Bundle\StudioBackendBundle\Dependency\Event\DependencyEvent;
 use Pimcore\Bundle\StudioBackendBundle\Dependency\Hydrator\DependencyHydratorInterface;
 use Pimcore\Bundle\StudioBackendBundle\Dependency\Repository\DependencyRepositoryInterface;
-use Pimcore\Bundle\StudioBackendBundle\Dependency\Request\DependencyParameters;
+use Pimcore\Bundle\StudioBackendBundle\Dependency\MappedParameter\DependencyParameters;
 use Pimcore\Bundle\StudioBackendBundle\Dependency\Response\Collection;
-use Pimcore\Bundle\StudioBackendBundle\Request\ElementParameters;
+use Pimcore\Bundle\StudioBackendBundle\MappedParameter\ElementParameters;
 use Pimcore\Bundle\StudioBackendBundle\Util\Traits\ElementProviderTrait;
 use Pimcore\Model\UserInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
