+++ src/Note/Service/FilterService.php
@@ -18,8 +18,8 @@ namespace Pimcore\Bundle\StudioBackendBundle\Note\Service;
 
 use Exception;
 use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidFilterException;
-use Pimcore\Bundle\StudioBackendBundle\Note\Request\NoteElement;
-use Pimcore\Bundle\StudioBackendBundle\Note\Request\NoteParameters;
+use Pimcore\Bundle\StudioBackendBundle\Note\MappedParameter\NoteElementParameters;
+use Pimcore\Bundle\StudioBackendBundle\Note\MappedParameter\NoteParameters;
 use Pimcore\Model\Element\Note\Listing as NoteListing;
 
 /**
@@ -80,7 +80,7 @@ final class FilterService implements FilterServiceInterface
         }
     }
 
-    public function applyElementFilter(NoteListing $list, NoteElement $noteElement): void
+    public function applyElementFilter(NoteListing $list, NoteElementParameters $noteElement): void
     {
         if ($noteElement->getId() && $noteElement->getType()) {
             $list->addConditionParam(
