+++ src/Note/Service/NoteServiceInterface.php
@@ -19,8 +19,8 @@ namespace Pimcore\Bundle\StudioBackendBundle\Note\Service;
 use Pimcore\Bundle\StudioBackendBundle\Exception\ElementNotFoundException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\ElementSavingFailedException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidFilterException;
-use Pimcore\Bundle\StudioBackendBundle\Note\Request\NoteElement;
-use Pimcore\Bundle\StudioBackendBundle\Note\Request\NoteParameters;
+use Pimcore\Bundle\StudioBackendBundle\Note\MappedParameter\NoteElementParameters;
+use Pimcore\Bundle\StudioBackendBundle\Note\MappedParameter\NoteParameters;
 use Pimcore\Bundle\StudioBackendBundle\Note\Response\Collection;
 use Pimcore\Bundle\StudioBackendBundle\Note\Schema\CreateNote;
 use Pimcore\Bundle\StudioBackendBundle\Note\Schema\Note;
@@ -33,12 +33,12 @@ interface NoteServiceInterface
     /**
      * @throws ElementSavingFailedException|ElementNotFoundException
      */
-    public function createNote(NoteElement $noteElement, CreateNote $createNote): Note;
+    public function createNote(NoteElementParameters $noteElement, CreateNote $createNote): Note;
 
     /**
      * @throws InvalidFilterException
      */
-    public function listNotes(NoteElement $noteElement, NoteParameters $parameters): Collection;
+    public function listNotes(NoteElementParameters $noteElement, NoteParameters $parameters): Collection;
 
     /**
      * @throws ElementNotFoundException
