+++ src/Version/MappedParameter/VersionCleanupParameters.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\Version\Request;
+namespace Pimcore\Bundle\StudioBackendBundle\Version\MappedParameter;
 
 /**
  * @internal
