+++ tests/Unit/Note/Service/FilterServiceTest.php
@@ -19,7 +19,7 @@ namespace Pimcore\Bundle\StudioBackendBundle\Tests\Unit\Note\Service;
 use Codeception\Test\Unit;
 use JsonException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\InvalidFilterException;
-use Pimcore\Bundle\StudioBackendBundle\Note\Request\NoteParameters;
+use Pimcore\Bundle\StudioBackendBundle\Note\MappedParameter\NoteParameters;
 use Pimcore\Bundle\StudioBackendBundle\Note\Service\FilterService;
 use Pimcore\Bundle\StudioBackendBundle\Note\Service\FilterServiceInterface;
 use Pimcore\Model\Element\Note\Listing as NoteListing;
